+++ src/main/java/io/vinum/capability/IBAC.java
@@ -2,10 +2,10 @@ package io.vinum.capability;
 
 public interface IBAC {
 	
-	public int getBACLevel();
-	public void setBACLevel(int BACLevel);
-	public void addBACLevel(int BACLevel);
-	public void removeBACLevel(int BACLevel);
+	public double getBACLevel();
+	public void setBACLevel(double BACLevel);
+	public void addBACLevel(double BACLevel);
+	public void removeBACLevel(double BACLevel);
 	public int getBACTicks();
 	public void setBACTicks(int BACTicks);
 	public void addBACTicks(int BACTicks);
